Project_4/src/task2/Book.java
package task2;

public class Book {

    private String name;
    private String authorName;
    private String publisher;
    private int publishingYear;
    private float price;

    public Book(String name, String authorName, String publisher, int publishingYear, float price) {
        this.name = name;
        this.authorName = authorName;
        this.publisher = publisher;
        this.publishingYear = publishingYear;
        this.price = price;
    }

    public Book() {
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getAuthorName() {
        return authorName;
    }

    public void setAuthorName(String authorName) {
        this.authorName = authorName;
    }

    public String getPublisher() {
        return publisher;
    }

    public void setPublisher(String publisher) {
        this.publisher = publisher;
    }

    public int getPublishingYear() {
        return publishingYear;
    }

    public void setPublishingYear(int publishingYear) {
        this.publishingYear = publishingYear;
    }

    public float getPrice() {
        return price;
    }

    public void setPrice(float price) {
        this.price = price;
    }

    @Override
    public String toString() {
        return "Book: " +
                "name='" + name + '\'' +
                ", authorName='" + authorName + '\'' +
                ", publisher='" + publisher + '\'' +
                ", publishingYear=" + publishingYear +
                ", price=" + price;
    }
}
